+++ Retargeting/Retargeting/WangHash.cpp
@@ -1,6 +1,6 @@
 #include "WangHash.h"
 
-uint32_t WangHash::generate(uint32_t seed) {
+uint64_t WangHash::generate(uint64_t seed) {
     seed = (seed ^ 61) ^ (seed >> 16);
     seed *= 9;
     seed = seed ^ (seed >> 4);
