+++ Retargeting/Retargeting/WangHash.h
@@ -6,6 +6,6 @@
 
 class WangHash : public RandomnessStrategy {
 public:
-	uint32_t generate(uint32_t seed);
+	uint64_t generate(uint64_t seed);
 	std::string getName();
 };
